scrub my project id and update readme for sed scripting to replace PROJECT and HISTORY_BUCKET

+++ examples/dataproc-persistent-history-server/init_actions/disable_history_servers.sh
@@ -1,3 +1,7 @@
-#!/usr/bin/env bash
-systemctl stop hadoop-mapreduce-historyserver
-systemctl stop spark-history-server
+#!/bin/bash
+set -euxo pipefail
+role="$(/usr/share/google/get_metadata_value attributes/dataproc-role)"
+if [[ "${role}" == 'Master' ]]; then
+  systemctl stop hadoop-mapreduce-historyserver
+  systemctl stop spark-history-server
+fi
